feat: added package manager `vite`
