+++ .env
@@ -1,0 +1,2 @@
+VITE_APP_SiteName = "EOE.LOL"
+VITE_APP_EOEFansKey = "25aac10cef164deca8c98a2b4763bdb5"
