--- config.js
@@ -1,2 +1,0 @@
-const siteName = "EOE.LOL"; //网站名称
-const EOEFansKey = "25aac10cef164deca8c98a2b4763bdb5"; // 请自行申请，你不可以直接使用 EOE.LOL 的 API Key。
