+++ legacy-video.html
@@ -7,13 +7,13 @@
     <meta name="viewport"
         content="width=device-width, initial-scale=1.0, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0" />
     <meta name="theme-color" content="#0ED242" />
-    <link rel="shortcut icon" href="./favicon.png">
+    <link rel="shortcut icon" href="/favicon.png">
     <title>Powered by OpenEOELOL/OpenFansite</title>
     <script src="config.js"></script>
-    <link rel="stylesheet" href="./src/water.min.css" />
-    <link rel="stylesheet" href="./src/material-icon.css" />
-    <link rel="stylesheet" href="./custom.css" />
-    <link rel="stylesheet" href="./src/main.css" />
+    <link rel="stylesheet" href="/src/water.min.css" />
+    <link rel="stylesheet" href="/src/material-icon.css" />
+    <link rel="stylesheet" href="/custom.css" />
+    <link rel="stylesheet" href="/src/main.css" />
 </head>
 
 <body>
@@ -42,7 +42,7 @@
                 <p class="infinite-scroll-last">没有更多内容了</p>
                 <p class="infinite-scroll-error">不能再加载了/也有可能是服务器错误<br><br>
                     <a href="https://b23.tv/av565443822" onclick="video('565443822');return false;"><img
-                            src="./src/av565443822.png" alt="危险派对封面" class="img" loading="lazy"></a>我劝你多看
+                            src="/assets/picture/av565443822.png" alt="危险派对封面" class="img" loading="lazy"></a>我劝你多看
                 </p>
             </div>
             <button class="view-more-button" id="videoList_viewmore">
@@ -51,138 +51,7 @@
         </div>
     </div>
     </div>
-
-    <script src="./src/masonry.pkgd.min.js"></script>
-    <script src="./src/infinite-scroll.pkgd.min.js"></script>
-    <script src="./src/webfont.js"></script>
-    <script src="./src/limit.js"></script>
-    <script>
-        let OpenWay = localStorage.getItem("open_bilibili"); // 如何打开哔哩哔哩？
-
-
-
-        // 视频打开逻辑
-        function video(videoID) {
-            let link;
-            if (OpenWay == "app") {
-                link = "bilibili://video/" + videoID;
-                window.location.href = link;
-            } else if (OpenWay == "uwp") {
-                link = "richasy-bili://play?video=" + videoID;
-            } else if (OpenWay == "website") {
-                link = "https://www.bilibili.com/video/av" + videoID;
-                window.open(link, "_newtab");
-            }
-        }
-
-        window.location.hash = "#videopage";
-
-        // 视频列表 部分脚本 开始 //
-        // (这段其实压根就不是我写的 参考 Masonry 和 InfiniteScroll 官网展示的 codepen)
-        let video__pageNumber = 1;
-
-        // 瀑布流插件初始化
-        let video__msnry = new Masonry(".videoList", {
-            itemSelector: ".videoCard",
-            columnWidth: ".videoList__col-sizer",
-            gutter: ".videoList__gutter-sizer",
-            percentPosition: true,
-            stagger: 30,
-            transitionDuration: 0,
-            // nicer reveal transition
-            visibleStyle: { transform: "translateY(0)", opacity: 1 },
-            hiddenStyle: { transform: "translateY(100px)", opacity: 0 },
-        });
-
-        // 无限滚动插件初始化
-        let video__infScroll = new InfiniteScroll(".videoList", {
-            path: function () {
-                //let video__pageNumber = this.pageIndex;
-                return `https://api.eoe.lol/api/${video__pageNumber}`;
-            },
-            responseBody: "json", // 响应体为 JSON 格式
-            outlayer: video__msnry,
-            status: "#videoStatus", // 加载状态
-            history: false, // 不展示历史
-            // prefill: true, // 预先加载
-            scrollThreshold: false, // 不需要滚动到底部加载
-            button: "#videoList_viewmore", // 展示更多按钮定义
-        });
-
-        //加载时的事件
-        video__infScroll.on("request", function () {
-            //console.log(video__pageNumber);
-            video__pageNumber++; //每加载一页就准备好下一页的页数，使用自增语法。
-        });
-
-        // 转换 HTML 字符串到元素，用代理元素。
-        var video__proxyElem = document.createElement("div");
-
-        video__infScroll.on("load", function (body) {
-            // 数据转入 HTML 字符串
-            var video__itemsHTML = body["data"]
-                .map(getVideoItemHTML)
-                .join("");
-            // 将 HTML 字符串转入到元素
-            video__proxyElem.innerHTML = video__itemsHTML;
-            // 添加元素物件
-            let video__items = video__proxyElem.querySelectorAll(".videoCard");
-            // imagesLoaded(video__items, function () {
-            //     video__infScroll.appendItems(video__items); //添加元素给无限滚动处理
-            //     video__msnry.appended(video__items); //添加元素给瀑布流插件处理
-            // });
-            video__infScroll.appendItems(video__items); //添加元素给无限滚动处理
-            video__msnry.appended(video__items); //添加元素给瀑布流插件处理
-
-            setTimeout(() => {
-                debounce(AgainLayout, 500, false); //加载七百毫秒后重新排列
-            }, 700);
-        });
-
-        // 加载一次先
-        video__infScroll.loadNextPage();
-
-        // 此处定义一下一个视频卡片物件的 HTML 代码
-        function getVideoItemHTML({ title, author, cover, av }) {
-            return `<a class="videoCard" href="https://bilibili.com/video/av${av}" onclick="video('${av}');return false;">
-    <img src = "${cover}" alt="${author}的视频封面" loading="lazy" onload="AgainLayout()" />
-    <div>
-        <div>${title}</div>
-        <div>[UP]${author}</div>
-    </div>
-</a>`;
-        }
-        // 视频列表 部分脚本 结束 //
-        function AgainLayout() {
-            video__msnry.layout();
-        }
-        // 窗口大小变化时重新整理卡片位置
-
-        window.addEventListener("resize", debounce(AgainLayout, 500, false));
-
-        // 设置 部分脚本 开始 //
-
-        if (OpenWay == null) {
-            localStorage.setItem("open_bilibili", "website");
-        }
-
-        // 设置 部分脚本 结束 //
-
-        // 配置应用 部分脚本 开始 //
-        document.title = siteName + " Legacy";
-
-        WebFont.load({
-            custom: {
-                families: ["LXGWFasmartGothic"],
-            },
-            classes: false,
-            active() {
-                document.body.style.fontFamily = "LXGWFasmartGothic";
-            },
-        });
-
-            // 配置应用 部分脚本 结束 //
-    </script>
+    <script type="module" src="./src/page/legacy-video/index.js"></script>
 </body>
 
 </html>
