+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "vite-project",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^4.0.0"
+  },
+  "dependencies": {
+    "a11y-dialog": "^7.5.2",
+    "infinite-scroll": "^4.0.1",
+    "masonry-layout": "^4.2.2",
+    "webfontloader": "^1.6.28"
+  }
+}
