+++ src/page/index/main.css
@@ -332,7 +332,7 @@ dialog::-webkit-scrollbar {
 .banner {
     width: 100%;
     height: 50%;
-    background-image: url("../example/example.png");
+    background-image: url("/example/example.png");
     background-size: cover;
     background-position: center;
     border-radius: var(--radius);
